Refactor: Split interface.py into protocol-specific modules
Split the 1,547-line interface.py file into a subpackage with
separate modules for each interface type:

- interface/debug.py: DebugHeader (SWD, JTAG, Tag-Connect)
- interface/usb.py: USBConnector (Type-C, Micro-B, Mini-B, Type-A)
- interface/i2c.py: I2CPullups with filtering
- interface/can.py: CANTransceiver with termination and ESD

The interface/__init__.py re-exports all public APIs, maintaining
backward compatibility with existing imports from blocks.interface.
